Extracting logic from QueuesTable to custom hook
